--- src/bootstrap/__init__.py
@@ -1,5 +1,0 @@
-# Copyright 2018 Open Source Foundries Limited.
-#
-# SPDX-License-Identifier: Apache-2.0
-
-# Empty file.
