+++ src/west/cmd/command.py
