+++ src/west/cmd/flash.py
@@ -4,8 +4,8 @@
 
 '''west "flash" command'''
 
-from .run_common import desc_common, add_parser_common, do_run_common
-from . import WestCommand
+from west.cmd.run_common import desc_common, add_parser_common, do_run_common
+from west.cmd.command import WestCommand
 
 
 class Flash(WestCommand):
