+++ src/west/runner/jlink.py
@@ -7,8 +7,8 @@
 import os
 import tempfile
 
-from .. import log
-from .core import ZephyrBinaryRunner, RunnerCaps, BuildConfiguration
+from west import log
+from west.runner.core import ZephyrBinaryRunner, RunnerCaps, BuildConfiguration
 
 DEFAULT_JLINK_GDB_PORT = 2331
