--- tests/west/project/__init__.py
@@ -1,1 +1,0 @@
-# Nothing here.
